projects/phase-05-project-copy/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.3.1
rank_bm25>=0.2.2
jieba>=0.42.1
rich>=13.7.0
python-dotenv>=1.0.0
pypdf>=3.17.4
unstructured[all-docs]>=0.12.0
